进货单优化

+++ src/components/ImageUpload/index.vue
@@ -49,6 +49,7 @@
 <script setup>
 import { getToken } from "@/utils/auth";
 import request from '@/utils/request';
+import Compressor from 'compressorjs';
 
 const props = defineProps({
   modelValue: [String, Object, Array],
@@ -60,7 +61,7 @@ const props = defineProps({
   // 大小限制(MB)
   fileSize: {
     type: Number,
-    default: 5,
+    default: 10,
   },
   // 文件类型, 例如['png', 'jpg', 'jpeg']
   fileType: {
@@ -117,6 +118,7 @@ watch(() => props.modelValue, val => {
   }
 },{ deep: true, immediate: true });
 
+
 // 上传前loading加载
 function handleBeforeUpload(file) {
   let isImg = false;
@@ -139,15 +141,37 @@ function handleBeforeUpload(file) {
     );
     return false;
   }
+  console.log("开始校验图片大小。。")
   if (props.fileSize) {
     const isLt = file.size / 1024 / 1024 < props.fileSize;
     if (!isLt) {
-      proxy.$modal.msgError(`上传头像图片大小不能超过 ${props.fileSize} MB!`);
+      proxy.$modal.alert(`上传图片大小不能超过 ${props.fileSize} MB!`);
       return false;
     }
   }
-  proxy.$modal.loading("正在上传图片，请稍候...");
-  number.value++;
+
+  return new Promise((resolve, reject) => {
+    new Compressor(file, {
+      quality: 0.6, // 压缩质量
+      success: (result) => { // 使用箭头函数
+        // 显示压缩后的图片
+        console.log('成功压缩图片:', file.size / 1024 / 1024, result.size / 1024 / 1024);
+        resolve(result); // 解析压缩后的文件
+      },
+      error: (err) => { // 同样使用箭头函数
+        console.error('压缩失败:', err.message);
+        reject(err); // 拒绝并返回错误
+      }
+    });
+  }).then((compressedFile) => {
+    proxy.$modal.loading("正在上传图片，请稍候...");
+    number.value++;
+    return compressedFile; // 返回压缩后的文件以便继续上传
+  }).catch((err) => {
+    proxy.$modal.closeLoading();
+    proxy.$modal.msgError("图片压缩失败，请重试！");
+    return Promise.reject(err); // 返回错误
+  });
 }
 
 // 文件个数超出
@@ -171,19 +195,38 @@ function handleUploadSuccess(res, file) {
 }
 
 function beforeDel(file) {
-  proxy.$modal.confirm('确认删除图片？').then(function() {
-    return handleDelete(file);
-  })
-  return false;
+  console.log("beforeDel:"+ file.name);
+  console.log(file.name.startsWith(baseUrl))
+  // 说明：这个方法会有2中情况被调用：
+  // 1.手动删除时调用，也就是手动删除已经上传成功的图片，已经上传成功的图片，都是服务器返回的图片。
+  //     服务器返回的图片的name都是以baseUrl开头的。所以如果是删除服务器图片，就要先弹窗确认。
+  //     防止误删。
+  // 2.自动调用。什么时候会自动调用这个方法呢？就是当上传文件不符合要求时会调用，比如文件过大。或者文件格式不对。或者
+  //     文件数量超出限制。这些都会自动调用这个方法。这个时候，就要直接删除，也就是直接返回true就行。不应该弹窗确认是否
+  //     删除。
+  if (file.name.startsWith(baseUrl)) {
+    proxy.$modal.confirm('确认删除图片？').then(function() {
+      return handleDelete(file);
+    })
+  } else {
+    return true;
+  }
+  
 }
-
 
 // 删除图片
 async function  handleDelete(file) {
+  console.log("handleDelete:"+ file.name);
+  console.log("fileList:", fileList);
+  console.log("number:", uploadList.value.length, number.value)
   const findex = fileList.value.map(f => f.name).indexOf(file.name);
   if (findex > -1 && uploadList.value.length === number.value) {
+    console.log("handleDelete2:"+ file.name);
     //删除请求
-    var deleteFilePath = [];
+    
+  }
+
+  var deleteFilePath = [];
     const filePath = file.name.indexOf(baseUrl) > -1 ? file.name.substring(baseUrl.length) : file.name;
     deleteFilePath.push(filePath);
     const response = await request({
@@ -192,21 +235,22 @@ async function  handleDelete(file) {
       data: deleteFilePath
     });
     
-    fileList.value.splice(findex, 1);
+    // fileList.value.splice(findex, 1);
     emit("update:modelValue", uploadList.value);
-    return false;
-  }
+    return true;
 }
 
 // 上传结束处理
 function uploadedSuccessfully() {
   if (number.value > 0 && uploadList.value.length === number.value) {
-    fileList.value = fileList.value.filter(f => f.url !== undefined).concat(uploadList.value);
+    
+  }
+
+  fileList.value = fileList.value.filter(f => f.url !== undefined).concat(uploadList.value);
     uploadList.value = [];
     number.value = 0;
     emit("update:modelValue", uploadList.value);
     proxy.$modal.closeLoading();
-  }
 }
 
 // 上传失败
